+++ CHANGES.md
@@ -1,10 +1,12 @@
-0.1.11
+0.2
 ====
 
 * Expose `Symbol` and `Nat` types from `GHC.TypeLits` by default.
 * Switch exported `(<>)` to be from `Data.Monoid` instead of Semigroup.
 * Expose `putByteString` and `putLByteString` monomorphic versions of `putStrLn` functions
 * Export `genericLength` and other generic list return functions.
+* Export `ExceptT`, `ReaderT`, and `StateT` constructors.
+* Export `NonEmpty` type and constructor for GHC 8.0.
 
 0.1.9
 ====